Add files via upload
In India, there are almost 40 million blind people among which 1.6 million are children. Blind people face great difficulty in walking independently. They have to depend on others in many aspects of their life. The Major problem is when they walk on the road. With a stick in hand, they cannot detect every obstacle that comes their way.
IoT-based Smart shoe system for the blind is a system made with the help of ultrasonic sensors paired to an Arduino NANO board. The Internet of Things is about making physical objects communicate with other objects or even humans. It is an enabling technology which has a rapid development and growth in the market.